requirements.txt
Flask==3.0.0
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv
Flask-Migrate
flask-livereload